refactoring VPackCreatorMethodUtils
